feat: :sparkles: form confirmation modal and customer backend logic

+++ backend/src/main/java/com/tienda/category/application/service/CategoryService.java
@@ -7,8 +7,8 @@ import com.tienda.category.domain.entity.Category;
 
 public interface CategoryService {
     Set<Category> findAll();
-    Optional<Category> findById(Integer id);
+    Optional<Category> findById(Long id);
     Category save(Category category);
-    Category update(Integer id, Category category);
-    Optional<Category> delete(Integer id);
+    Category update(Long id, Category category);
+    Optional<Category> delete(Long id);
 }
